Minor project tweaks and fixes.
- added CPU_Z80_DEPENDENCIES_H to remaining build configurations.
- fixed start wrong address for JSW2+ start room.
- fixed 64-bit build warning with SoundOut call.
- fixed typo in test snapshot filename.
- removed unused Z submodule.

+++ SpectrumGame.cpp
@@ -64,7 +64,7 @@ void SpectrumGame::SetMap(const std::vector<MapRoom> &map)
 			if (td.pGame->IsActiveTile(td))
 			{
 				if (active && (address & 0x00ff) == 0x00fe)
-					SoundOut(value, td.frame_cycles + td.z80.cycles);
+					SoundOut(value, static_cast<int>(td.frame_cycles + td.z80.cycles));
 			}
 		};
 		tile.z80.int_data = [](void *context) -> zuint32 {
